[Refactor] 메뉴 로딩(공휴일 포함) 로직을 UseCase로 이동합니다
