+++ app/src/main/java/com/eatssu/android/domain/model/MenuLoadResult.kt
@@ -1,0 +1,8 @@
+package com.eatssu.android.domain.model
+
+import com.eatssu.common.enums.Restaurant
+
+data class MenuLoadResult(
+    val menuMap: Map<Restaurant, List<Menu>>,
+    val publicHolidayName: String?,
+)
